use longs
git-svn-id: svn://svn.coreboot.org/coreboot/trunk@1055 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/arch/ppc/include/timer.h
@@ -4,9 +4,9 @@
 #ifndef _TIMER_H
 #define __TIMER_H
 
-unsigned get_hz(void);
-unsigned ticks_since_boot(void);
-void sleep_ticks(unsigned);
+unsigned long get_hz(void);
+unsigned long ticks_since_boot(void);
+void sleep_ticks(unsigned long);
 void udelay(int);
 
 #endif
